Separated components into different .jsx files
