--- react-client/src/components/List.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import ListItem from './ListItem.jsx';
-
-const List = (props) => (
-  <div>
-    <h4> List Component </h4>
-    There are { props.items.length } items.
-    { props.items.map(item => <ListItem item={item}/>)}
-  </div>
-)
-
-export default List;
