--- react-client/src/components/ListItem.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const ListItem = (props) => (
-  <div>
-    { props.item.description }
-  </div>
-)
-
-export default ListItem;
