+++ react-client/src/index.jsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import axios from 'axios';
-import List from './components/List.jsx';
+import Navigation from './components/Navigation.jsx';
+import CategoryView from './components/CategoryView.jsx';
+import CourseDetailView from './components/CourseDetailView.jsx';
+import LoginModal from './components/LoginModal.jsx';
 
 class App extends React.Component {
   constructor(props) {
@@ -32,7 +35,7 @@ class App extends React.Component {
         {
           _id: 1,
           name: 'johncrogers',
-          passwrod: '1234',
+          password: '1234',
           coursesUpvoted: [],
         },
       ],
@@ -56,106 +59,6 @@ class App extends React.Component {
         <CategoryView categories={this.state.categories} />
         <CourseDetailView course={this.state.categories[0].courses[0]} />
         <LoginModal />
-      </div>
-    );
-  }
-}
-
-class Navigation extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      selectedCategory: '',
-      modalTriggered: false,
-    };
-  }
-  render() {
-    return (
-      <nav>
-        <a href="">Home</a>
-        <select placeholder="select a catergory">
-          {this.props.categories.map(category => <option>{category.name}</option>)}
-        </select>
-        <button>Sign In</button>
-        <button>Sign Out</button>
-        <button>Log Out</button>
-      </nav>
-    );
-  }
-}
-
-const CategoryView = props => (
-  <div>
-    <h3>Category: </h3>
-    <CategoryViewCourse course={props.categories[0].courses[0]} />
-  </div>
-);
-
-class CategoryViewCourse extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      upVoteCount: 0,
-      // downVoteCount = 0
-    };
-  }
-  render() {
-    return (
-      <div className="make me a card">
-        <h3>Course Title: {this.props.course.name}</h3>
-        <span>Price: ${this.props.course.description.price}</span>
-        <span>Instructor: {this.props.course.description.instructor}</span>
-        <span>Upvote Count: {this.state.upVoteCount}</span>
-        <p>{this.props.course.description.text}</p>
-        // Display course detail view onClick
-      </div>
-    );
-  }
-}
-
-class CourseDetailView extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-  render() {
-    return (
-      <div>
-        <div>
-          <button>Back</button>
-          <h3>Course Title: {this.props.course.name}</h3>
-          <span>Hosted: {this.props.course.courseUrl}</span>
-          <span>Price: {this.props.course.description.price}$ </span>
-        </div>
-        <div>
-          <iframe src={this.props.course.description.videoUrl} />
-          <p>Decription: {this.props.course.description.text}</p>
-        </div>
-      </div>
-    );
-  }
-}
-
-class LoginModal extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      email: '',
-      password: '',
-    };
-  }
-
-  render() {
-    return (
-      <div>
-        <button>Close</button>
-        <form>
-          <label>Username</label>
-          <input type="email" />
-          <label>Password</label>
-          <input type="password" />
-        </form>
-        <button>Sign In</button>
-        <button>Sign Up</button>
       </div>
     );
   }
